audio/openai_stt_streamlit/pyproject.toml
[project]
name = "openai-tts-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
